More refactoring.
Adds the trigger_pulled component that will be added when
space is pressed - ah, but what if we already have that component?
Should we do this in an entirely different system? Whaaat?

I think the code crashes as it is written now. Fudge.
