+++ src/components/control/mod.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::Component;
+
+#[derive(Component, Clone)]
+pub struct TriggerPulled {}
