+++ src/main.rs
@@ -6,11 +6,12 @@ use bevy_prototype_lyon::plugin::ShapePlugin;
 use bevy_rand::plugin::EntropyPlugin;
 use bevy_xpbd_2d::prelude::*;
 use big_brain::{BigBrainPlugin, BigBrainSet};
-use boids::{boid_steering, BoidDirection, BoidStuff, find_prey_action_system, hunger_scorer_system, hunger_system, hunt_prey_action_system, quad_boid_flocking, spawn_boids};
 use systems::*;
-use crate::boids::{attack_and_eat_action_system, Hunger, HuntTarget};
-use crate::components::{PlayerControl, Health, QuadCoord, QuadStore};
+use crate::components::{Health, PlayerControl, QuadCoord, QuadStore};
 use rand_chacha::ChaCha8Rng;
+use boids::ai::{attack_and_eat_action_system, find_prey_action_system, Hunger, hunger_scorer_system, hunger_system, hunt_prey_action_system, HuntTarget};
+use boids::components::{BoidDirection, BoidStuff};
+use boids::systems::{boid_steering, quad_boid_flocking, spawn_boids};
 use systems::camera::camera_follow;
 use systems::input::{add_mouse_aim_line, draw_mouse_aim, keyboard_input, mouse_look, mouse_position};
 use systems::movement::{linear_velocity_control_boid, linear_velocity_control_player};
